agents/zazu/scripts/requirements.txt
google-api-python-client
google-auth-oauthlib
feedparser
